move trace-dump into trace
