+++ rust/.gitignore
@@ -4,3 +4,6 @@
 *.rs.bk
 target
 build
+
+# Editor configs
+.idea/
