+++ rust/core-lib/Cargo.toml
@@ -19,7 +19,6 @@ regex = "1.0"
 memchr = "2.0.1"
 
 xi-trace = { path = "../trace", version = "0.1.0" }
-xi-trace-dump = { path = "../trace-dump", version = "0.1.0" }
 xi-rope = { path = "../rope", version = "0.3" }
 xi-unicode = { path = "../unicode", version = "0.1.0" }
 xi-rpc = { path = "../rpc", version = "0.2.0" }
